backend/positioning/requirements.txt
prisma==0.13.*
fastapi==0.111.0
uvicorn[standard]==0.30.*
datamodel-code-generator==0.25.*
jsonschema==4.22.*
python-dateutil==2.9.*
geojson==3.1.*
shapely==2.0.*
pyturf==0.6.*